+++ src/lib.rs
@@ -778,34 +778,60 @@ impl<M: Model> InnerView<M> {
 /// for most applications.
 ///
 /// # Example
+///
 /// ```
 /// let options = nutmeg::Options::default()
 ///     .progress_enabled(false); // Don't draw bars, only print.
 /// ```
-#[non_exhaustive]
+///
+/// Options can be constructed as a static or constant value, using [Options::new].
+///
+/// ```
+/// use std::time::Duration;
+/// use nutmeg::Options;
+///
+/// static NUTMEG_OPTIONS: Options = Options::new()
+///     .update_interval(Duration::from_millis(100))
+///     .progress_enabled(true)
+///     .destination(nutmeg::Destination::Stderr);
+/// ```
 #[derive(Debug, Clone)]
 pub struct Options {
     /// Target interval to repaint the progress bar.
-    pub update_interval: Duration,
+    update_interval: Duration,
 
     /// How long to wait after printing output before drawing the progress bar again.
-    pub print_holdoff: Duration,
+    print_holdoff: Duration,
 
     /// Is the progress bar drawn at all?
-    pub progress_enabled: bool,
+    progress_enabled: bool,
 
     /// Use a fake clock for testing.
-    pub fake_clock: bool,
+    fake_clock: bool,
 
     /// Write progress and messages to stdout, stderr, or a capture buffer for tests?
-    pub destination: Destination,
+    destination: Destination,
 }
 
 impl Options {
+    /// Return some reasonable default options.
+    ///
+    /// The update interval and print holdoff are 100ms, the progress bar is enabled,
+    /// and output is sent to stdout.
+    pub const fn new() -> Options {
+        Options {
+            update_interval: Duration::from_millis(100),
+            print_holdoff: Duration::from_millis(100),
+            progress_enabled: true,
+            fake_clock: false,
+            destination: Destination::Stdout,
+        }
+    }
+
     /// Set whether the progress bar will be drawn.
     ///
     /// By default it is drawn, except that this value will be ignored by [View::new] if stdout is not a terminal.
-    pub fn progress_enabled(self, progress_enabled: bool) -> Options {
+    pub const fn progress_enabled(self, progress_enabled: bool) -> Options {
         Options {
             progress_enabled,
             ..self
@@ -815,7 +841,7 @@ impl Options {
     /// Set the minimal interval to repaint the progress bar.
     ///
     /// `Duration::ZERO` can be used to cause the bar to repaint on every update.
-    pub fn update_interval(self, update_interval: Duration) -> Options {
+    pub const fn update_interval(self, update_interval: Duration) -> Options {
         Options {
             update_interval,
             ..self
@@ -829,7 +855,7 @@ impl Options {
     /// repeatedly printing messages at short intervals.
     ///
     /// `Duration::ZERO` can be used to disable this behavior.
-    pub fn print_holdoff(self, print_holdoff: Duration) -> Options {
+    pub const fn print_holdoff(self, print_holdoff: Duration) -> Options {
         Options {
             print_holdoff,
             ..self
@@ -846,7 +872,7 @@ impl Options {
     ///
     /// If this is enabled the fake clock can be updated with
     /// [View::set_fake_clock].
-    pub fn fake_clock(self, fake_clock: bool) -> Options {
+    pub const fn fake_clock(self, fake_clock: bool) -> Options {
         Options { fake_clock, ..self }
     }
 
@@ -857,7 +883,7 @@ impl Options {
     /// [Destination::Stderr] may be useful for programs that expect stdout to be redirected
     /// to a file and that want to draw progress output that is not captured by the
     /// redirection.
-    pub fn destination(self, destination: Destination) -> Options {
+    pub const fn destination(self, destination: Destination) -> Options {
         Options {
             destination,
             ..self
@@ -868,16 +894,9 @@ impl Options {
 impl Default for Options {
     /// Create default reasonable view options.
     ///
-    /// The update interval and print holdoff are 100ms, the progress bar is enabled,
-    /// and output is sent to stdout.
+    /// This is the same as [Options::new].
     fn default() -> Options {
-        Options {
-            update_interval: Duration::from_millis(100),
-            print_holdoff: Duration::from_millis(100),
-            progress_enabled: true,
-            fake_clock: false,
-            destination: Destination::Stdout,
-        }
+        Options::new()
     }
 }